==> 9_DE_analysis.Rmd <==
---
title: "DE_analysis"
output: html_notebook
---

```{r}
#Load in packages
library(ggplot2) 
library(fitdistrplus) 
library(MASS) 
library(tidyr) 
library(gdata)
library(Seurat)
library(data.table)
library(EnvStats)
library(purrr)
library(dplyr)
library(sn)
library(matrixStats)
library(fmsb)
library(lme4)

#####################################
### Format and transform the data ###
#####################################

Micro1_for_DE_test <- readRDS("~/Documents/10X_data/Parallel_LOAD_Normal_RNA_final/Micro1_for_DE_test.rds")
allgenes_JG <- t(as.data.frame(Micro1_for_DE_test@assays$RNA@counts))
allgenes_JG <- as.data.frame(allgenes_JG)
ngenes <- ncol(allgenes_JG)
allgenes_JG$orig.ident <- Micro1_for_DE_test@meta.data$orig.ident
allgenes_JG$sampID <- Micro1_for_DE_test@meta.data$sampID
allgenes_JG$wellKey <- paste(Micro1_for_DE_test@meta.data$sampID, rownames(Micro1_for_DE_test@meta.data), sep = "_")
rownames(allgenes_JG) <- allgenes_JG$wellKey
allgenes_JG$cell.type.number.proportion <- Micro1_for_DE_test@meta.data$cell.type.number.proportion
allgenes_JG <- allgenes_JG[,c(ngenes+3,ngenes+2,ngenes+1,ngenes+4,1:ngenes)]
genecounts_JG <- as.matrix(t(allgenes_JG[,c(-1,-2,-3,-4)]))
genecounts_JG <- log2(genecounts_JG + 1) #log2 transform
coldata_JG <- allgenes_JG[,1:4]
coldata_JG$orig.ident <- as.factor(coldata_JG$orig.ident)
nrow(genecounts_JG) #36588 genes

#####################################
###        Filter the data        ###
#####################################

#Calculate %cells expressing each gene in each group
LOAD.cells <- as.matrix(t(allgenes_JG[allgenes_JG$orig.ident == 'LOAD', c(-1,-2,-3,-4)]))
ncol(LOAD.cells) #4717
Normal.cells <- as.matrix(t(allgenes_JG[allgenes_JG$orig.ident == 'Normal', c(-1,-2,-3,-4)]))
ncol(Normal.cells) #4569

PercentAbove <- function(x, threshold){
  return(length(x = x[x > threshold]) / length(x = x))
}

pct.exp.LOAD <- apply(X = LOAD.cells, MARGIN = 1, FUN = PercentAbove, threshold = 0)
pct.exp.Normal <- apply(X = Normal.cells, MARGIN = 1, FUN = PercentAbove, threshold = 0)

# Filter out genes expressed in <10% of cells in 1 group 
alpha.min <- pmax(pct.exp.LOAD, pct.exp.Normal)
genes.to.keep <- names(x = which(x = alpha.min >= 0.1)) #6397 genes to keep
genecounts_JG <- genecounts_JG[rownames(genecounts_JG)%in%genes.to.keep,]

#####################################
###             Run MAST          ###
#####################################

genecounts_JG <- genecounts_JG[,rownames(coldata_JG)] #match order of data with metadata

fData_JG <- data.frame(primerid=rownames(genecounts_JG)) #feature data
sca_JG <- suppressMessages(MAST::FromMatrix(exprsArray=genecounts_JG, cData=coldata_JG, fData=fData_JG)) #makes a single-cell assay with expression data, metadata (e.g. Control_1_cell_1), and feature data
cdr2_JG <- colSums(SummarizedExperiment::assay(sca_JG)>0) #compute cellular detection rate (cdr)
SummarizedExperiment::colData(sca_JG)$ngeneson <- scale(cdr2_JG) #adds centered and scaled cdr to listData in colData of sca object
cond <-factor(SummarizedExperiment::colData(sca_JG)$orig.ident)
cond <-relevel(cond,"Normal") # set the reference level of orig.ident condition to be Normal
SummarizedExperiment::colData(sca_JG)$orig.ident <- cond
SummarizedExperiment::colData(sca_JG)$sampID <- factor(SummarizedExperiment::colData(sca_JG)$sampID) #set sampID to factor

zlmCond_JG_prop <- suppressMessages(MAST::zlm(~ orig.ident + ngeneson + cell.type.number.proportion + (1 | sampID), sca_JG, method='glmer',ebayes = F, fitArgsD = list(nAGQ = 0), strictConvergence = FALSE)) #runs zero-inflated regression fitting a generalized linear mixed-effects model.

colnames(coef(zlmCond_JG_prop, 'D')) #check names of modeled coefficients

# Generate summary of results with likelihood ratio test
summaryCond_JG_prop <- suppressMessages(MAST::summary(zlmCond_JG_prop,
                                                      doLRT='orig.identLOAD'))

#format the results
summaryDt_JG_prop <- summaryCond_JG_prop$datatable

fcHurdle_JG_prop <- merge(summaryDt_JG_prop[contrast=='orig.identLOAD' & component=='H',.(primerid, `Pr(>Chisq)`)], 
                          summaryDt_JG_prop[contrast=='orig.identLOAD' & component=='logFC', .(primerid, coef, ci.hi, ci.lo)], 
                          by='primerid')

fcHurdle_JG_fdr_prop <- fcHurdle_JG_prop[,fdr:=p.adjust(`Pr(>Chisq)`, 'fdr')] #add fdr adjusted p-values

#Add % cells from each group expressing each gene
pct.cells.LOAD <- as.data.frame(pct.exp.LOAD)
pct.cells.LOAD <- subset(pct.cells.LOAD, rownames(pct.cells.LOAD)%in%fcHurdle_JG_fdr_prop$primerid)
pct.cells.LOAD <- tibble::rownames_to_column(pct.cells.LOAD, var = "primerid")
pct.cells.Normal <- as.data.frame(pct.exp.Normal)
pct.cells.Normal <- subset(pct.cells.Normal, rownames(pct.cells.Normal)%in%fcHurdle_JG_fdr_prop$primerid)
pct.cells.Normal <- tibble::rownames_to_column(pct.cells.Normal, var = "primerid")
fcHurdle_JG_fdr_prop_pct.cells <- merge(fcHurdle_JG_fdr_prop, pct.cells.LOAD, by.x = 'primerid', by.y = 'primerid')
fcHurdle_JG_fdr_prop_pct.cells <- merge(fcHurdle_JG_fdr_prop_pct.cells, pct.cells.Normal, by.x = 'primerid', by.y = 'primerid')

fcHurdle_JG_prop_na_omit <- stats::na.omit(as.data.frame(fcHurdle_JG_fdr_prop_pct.cells)) #2 genes omitted - NA values occur from convergence failures

hist(fcHurdle_JG_prop_na_omit$fdr) #examine distribution of adjusted p-values
hist(fcHurdle_JG_prop_na_omit$coef, breaks = 25) #examine distribution of logFC
fcHurdleSig_prop_JG <- subset(fcHurdle_JG_prop_na_omit, subset = fdr<0.05) #determine number of significant genes


############################################################################################
### For datasets of >30,000 cells per group, replace above formatting with the following ###
############################################################################################

Oligo <- readRDS('/gpfs/fs1/data/chiba-faleklab/Parallel_ADRC_RNA_final/objects/Oligo_for_DE.rds')

DefaultAssay(Oligo) <- "RNA"
Oligo

#Remove genes that are not expressed
counts <- Oligo@assays$RNA@counts
nrow(counts) #36588
genes.use <- rowSums(counts) > 0
genes.use <- as.data.frame(genes.use)
head(genes.use)
genes.use <- subset(genes.use, genes.use=='TRUE')
genes.use <- as.vector(rownames(genes.use))
head(genes.use)
length(genes.use) #31720

Oligo <- subset(Oligo, features=genes.use)
counts.subset <-Oligo@assays$RNA@counts
nrow(counts.subset) #31720

#split matrix into submatrices, format,  log transform, then re-merge
Oligo1 <- Oligo[,1:45000]
Oligo2 <- Oligo[,45001:83945]

allgenes_JG <- t(as.data.frame(Oligo1@assays$RNA@counts))
allgenes_JG <- as.data.frame(allgenes_JG)
ngenes <- ncol(allgenes_JG)
allgenes_JG$orig.ident <- Oligo1@meta.data$orig.ident
allgenes_JG$sampID <- Oligo1@meta.data$sampID
allgenes_JG$Sex <- Oligo1@meta.data$Sex
allgenes_JG$age <- Oligo1@meta.data$age
allgenes_JG$PMI <- Oligo1@meta.data$PMI
allgenes_JG$seq.sat <- Oligo1@meta.data$seq.sat
allgenes_JG$wellKey <- paste(Oligo1@meta.data$sampID, rownames(Oligo1@meta.data), sep = "_")
rownames(allgenes_JG) <- allgenes_JG$wellKey
allgenes_JG$cell.type.number.proportion <- Oligo1@meta.data$cell.type.number.proportion
allgenes_JG <- allgenes_JG[,c(ngenes+7,ngenes+2,ngenes+1,ngenes+3,ngenes+4,ngenes+5,ngenes+6,ngenes+8,1:ngenes)]
genecounts_JG <- as.matrix(t(allgenes_JG[,c(-1,-2,-3,-4,-5,-6,-7,-8)]))
genecounts_JG <- log2(genecounts_JG + 1) #log2 transform
nrow(genecounts_JG) #31720 genes

allgenes_JG2 <- t(as.data.frame(Oligo2@assays$RNA@counts))
allgenes_JG2 <- as.data.frame(allgenes_JG2)
ngenes <- ncol(allgenes_JG2)
allgenes_JG2$orig.ident <- Oligo2@meta.data$orig.ident
allgenes_JG2$sampID <- Oligo2@meta.data$sampID
allgenes_JG2$Sex <- Oligo2@meta.data$Sex
allgenes_JG2$age <- Oligo2@meta.data$age
allgenes_JG2$PMI <- Oligo2@meta.data$PMI
allgenes_JG2$seq.sat <- Oligo2@meta.data$seq.sat
allgenes_JG2$wellKey <- paste(Oligo2@meta.data$sampID, rownames(Oligo2@meta.data), sep = "_")
rownames(allgenes_JG2) <- allgenes_JG2$wellKey
allgenes_JG2$cell.type.number.proportion <- Oligo2@meta.data$cell.type.number.proportion
allgenes_JG2 <- allgenes_JG2[,c(ngenes+7,ngenes+2,ngenes+1,ngenes+3,ngenes+4,ngenes+5,ngenes+6,ngenes+8,1:ngenes)]
genecounts_JG2 <- as.matrix(t(allgenes_JG2[,c(-1,-2,-3,-4,-5,-6,-7,-8)]))
genecounts_JG2 <- log2(genecounts_JG2 + 1) #log2 transform
nrow(genecounts_JG2) #31720 genes

genecounts_JG <- cbind(genecounts_JG, genecounts_JG2)
head(genecounts_JG)
length(rownames(genecounts_JG)) #31720
length(colnames(genecounts_JG)) #83945

allgenes_JG <- bind_rows(allgenes_JG, allgenes_JG2)
head(allgenes_JG)
length(rownames(allgenes_JG)) #83945
length(colnames(genecounts_JG)) #31728

coldata_JG <- allgenes_JG[,1:8]
coldata_JG$orig.ident <- as.factor(coldata_JG$orig.ident)
coldata_JG$Sex <- as.factor(coldata_JG$Sex)
```
